refactor(bindgen): move stream and future functions out of cstate
