+++ crates/js-component-bindgen/src/intrinsics/component.rs
@@ -2,8 +2,6 @@
 
 use std::fmt::Write as _;
 
-use crate::intrinsics::p3::async_future::AsyncFutureIntrinsic;
-use crate::intrinsics::p3::async_stream::AsyncStreamIntrinsic;
 use crate::intrinsics::p3::waitable::WaitableIntrinsic;
 use crate::intrinsics::{Intrinsic, RenderIntrinsicsArgs};
 use crate::source::Source;
@@ -122,26 +120,10 @@ impl ComponentIntrinsic {
                 let component_async_state_class = self.name();
                 let debug_log_fn = render_args.require_intrinsic(Intrinsic::DebugLog);
                 let rep_table_class = render_args.require_intrinsic(Intrinsic::RepTableClass);
-                let internal_stream_class =
-                    render_args.require_intrinsic(AsyncStreamIntrinsic::InternalStreamClass);
-                let global_stream_map =
-                    render_args.require_intrinsic(AsyncStreamIntrinsic::GlobalStreamMap);
-                let global_stream_table_map =
-                    render_args.require_intrinsic(AsyncStreamIntrinsic::GlobalStreamTableMap);
-                let internal_future_class =
-                    render_args.require_intrinsic(AsyncFutureIntrinsic::InternalFutureClass);
-                let global_future_map =
-                    render_args.require_intrinsic(AsyncFutureIntrinsic::GlobalFutureMap);
-                let global_future_table_map =
-                    render_args.require_intrinsic(AsyncFutureIntrinsic::GlobalFutureTableMap);
                 let waitable_class =
                     render_args.require_intrinsic(WaitableIntrinsic::WaitableClass);
-                let get_or_create_async_state_fn =
-                    render_args.require_intrinsic(Self::GetOrCreateAsyncState);
                 let promise_with_resolvers_fn =
                     render_args.require_intrinsic(Intrinsic::PromiseWithResolversPonyfill);
-                let stream_readable_end_class =
-                    render_args.require_intrinsic(AsyncStreamIntrinsic::StreamReadableEndClass);
                 let runtime_error_class =
                     render_args.require_intrinsic(Intrinsic::WebAssemblyRuntimeError);
 
@@ -602,375 +584,9 @@ impl ComponentIntrinsic {
                                 : {component_async_state_class}.TickResult.DONE;
                         }}
 
-                        addStreamEndToTable(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#addStreamEnd()] args', args);
-                            const {{ tableIdx, streamEnd }} = args;
-                            if (typeof streamEnd === 'number') {{ throw new Error("INSERTING BAD STREAMEND"); }}
-
-                            let {{ table, componentIdx }} = {global_stream_table_map}[tableIdx];
-                            if (componentIdx === undefined || !table) {{
-                                throw new Error(`invalid global stream table state for table [${{tableIdx}}]`);
-                            }}
-
-                            const handle = table.insert(streamEnd);
-                            streamEnd.setHandle(handle);
-                            streamEnd.setStreamTableIdx(tableIdx);
-
-                            const cstate = {get_or_create_async_state_fn}(componentIdx);
-                            const waitableIdx = cstate.handles.insert(streamEnd);
-                            streamEnd.setWaitableIdx(waitableIdx);
-
-                            {debug_log_fn}('[{component_async_state_class}#addStreamEnd()] added stream end', {{
-                                tableIdx,
-                                table,
-                                handle,
-                                streamEnd,
-                                destComponentIdx: componentIdx,
-                            }});
-
-                            return {{ handle, waitableIdx }};
-                        }}
-
                         createWaitable(args) {{
                             return new {waitable_class}({{ target: args?.target, }});
                         }}
-
-                        createReadableStreamEnd(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#createStreamEnd()] args', args);
-                            const {{ tableIdx, elemMeta, hostInjectFn }} = args;
-
-                            const {{ table: localStreamTable, componentIdx }} = {global_stream_table_map}[tableIdx];
-                            if (!localStreamTable) {{
-                                throw new Error(`missing global stream table lookup for table [${{tableIdx}}] while creating stream`);
-                            }}
-                            if (componentIdx !== this.#componentIdx) {{
-                                throw new Error('component idx mismatch while creating stream');
-                            }}
-
-                            const waitable = this.createWaitable();
-                            const streamEnd = new {stream_readable_end_class}({{
-                                tableIdx,
-                                elemMeta,
-                                hostInjectFn,
-                                pendingBufferMeta: {{}},
-                                target: `stream read end (lowered, @init)`,
-                                waitable,
-                            }});
-
-                            streamEnd.setWaitableIdx(this.handles.insert(streamEnd));
-                            streamEnd.setHandle(localStreamTable.insert(streamEnd));
-                            if (streamEnd.streamTableIdx() !== tableIdx) {{
-                                throw new Error("unexpectedly mismatched stream table");
-                            }}
-                            const streamEndWaitableIdx = streamEnd.waitableIdx();
-                            const streamEndHandle = streamEnd.handle();
-                            waitable.setTarget(`waitable for stream read end (lowered, waitable [${{streamEndWaitableIdx}}])`);
-                            streamEnd.setTarget(`stream read end (lowered, waitable [${{streamEndWaitableIdx}}])`);
-
-                            return {{
-                                waitableIdx: streamEndWaitableIdx,
-                                handle: streamEndHandle,
-                                streamEnd,
-                            }};
-                        }}
-
-                        createStream(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#createStream()] args', args);
-                            const {{ tableIdx, elemMeta, hostInjectFn }} = args;
-                            if (tableIdx === undefined) {{ throw new Error("missing table idx while adding stream"); }}
-                            if (elemMeta === undefined) {{ throw new Error("missing element metadata while adding stream"); }}
-
-                            const {{ table: localStreamTable, componentIdx }} = {global_stream_table_map}[tableIdx];
-                            if (!localStreamTable) {{
-                                throw new Error(`missing global stream table lookup for table [${{tableIdx}}] while creating stream`);
-                            }}
-                            if (componentIdx !== this.#componentIdx) {{
-                                throw new Error('component idx mismatch while creating stream');
-                            }}
-
-                            const readWaitable = this.createWaitable();
-                            const writeWaitable = this.createWaitable();
-
-                            const stream = new {internal_stream_class}({{
-                                tableIdx,
-                                elemMeta,
-                                readWaitable,
-                                writeWaitable,
-                                hostInjectFn,
-                            }});
-                            stream.setGlobalStreamMapRep({global_stream_map}.insert(stream));
-
-                            const writeEnd = stream.writeEnd();
-                            writeEnd.setWaitableIdx(this.handles.insert(writeEnd));
-                            writeEnd.setHandle(localStreamTable.insert(writeEnd));
-                            if (writeEnd.streamTableIdx() !== tableIdx) {{ throw new Error("unexpectedly mismatched stream table"); }}
-
-                            const writeEndWaitableIdx = writeEnd.waitableIdx();
-                            const writeEndHandle = writeEnd.handle();
-                            writeWaitable.setTarget(`waitable for stream write end (waitable [${{writeEndWaitableIdx}}])`);
-                            writeEnd.setTarget(`stream write end (waitable [${{writeEndWaitableIdx}}])`);
-
-                            const readEnd = stream.readEnd();
-                            readEnd.setWaitableIdx(this.handles.insert(readEnd));
-                            readEnd.setHandle(localStreamTable.insert(readEnd));
-                            if (readEnd.streamTableIdx() !== tableIdx) {{ throw new Error("unexpectedly mismatched stream table"); }}
-
-                            const readEndWaitableIdx = readEnd.waitableIdx();
-                            const readEndHandle = readEnd.handle();
-                            readWaitable.setTarget(`waitable for read end (waitable [${{readEndWaitableIdx}}])`);
-                            readEnd.setTarget(`stream read end (waitable [${{readEndWaitableIdx}}])`);
-
-                            return {{
-                                writeEnd,
-                                writeEndWaitableIdx,
-                                writeEndHandle,
-                                readEndWaitableIdx,
-                                readEndHandle,
-                                readEnd,
-                            }};
-                        }}
-
-                        getStreamEnd(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#getStreamEnd()] args', args);
-                            const {{ tableIdx, streamEndHandle, streamEndWaitableIdx }} = args;
-                            if (tableIdx === undefined) {{
-                                throw new Error('missing table idx while getting stream end');
-                            }}
-
-                            const {{ table, componentIdx }} = {global_stream_table_map}[tableIdx];
-                            const cstate = {get_or_create_async_state_fn}(componentIdx);
-
-                            let streamEnd;
-                            if (streamEndWaitableIdx !== undefined) {{
-                                streamEnd = cstate.handles.get(streamEndWaitableIdx);
-                            }} else if (streamEndHandle !== undefined) {{
-                                if (!table) {{ throw new Error(`missing/invalid table [${{tableIdx}}] while getting stream end`); }}
-                                streamEnd = table.get(streamEndHandle);
-                            }} else {{
-                                throw new TypeError("must specify either waitable idx or handle to retrieve stream");
-                            }}
-
-                            if (!streamEnd) {{
-                                throw new Error(`missing stream end (tableIdx [${{tableIdx}}], handle [${{streamEndHandle}}], waitableIdx [${{streamEndWaitableIdx}}])`);
-                            }}
-                            if (tableIdx && streamEnd.streamTableIdx() !== tableIdx) {{
-                                throw new Error(`stream end table idx [${{streamEnd.streamTableIdx()}}] does not match [${{tableIdx}}]`);
-                            }}
-
-                            return streamEnd;
-                        }}
-
-                        deleteStreamEnd(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#deleteStreamEnd()] args', args);
-                            const {{ tableIdx, streamEndWaitableIdx }} = args;
-                            if (tableIdx === undefined) {{ throw new Error("missing table idx while removing stream end"); }}
-                            if (streamEndWaitableIdx === undefined) {{ throw new Error("missing stream idx while removing stream end"); }}
-
-                            const {{ table, componentIdx }} = {global_stream_table_map}[tableIdx];
-                            const cstate = {get_or_create_async_state_fn}(componentIdx);
-
-                            const streamEnd = cstate.handles.get(streamEndWaitableIdx);
-                            if (!streamEnd) {{
-                                throw new Error(`missing stream end [${{streamEndWaitableIdx}}] in component handles while deleting stream`);
-                            }}
-                            if (streamEnd.streamTableIdx() !== tableIdx) {{
-                                throw new Error(`stream end table idx [${{streamEnd.streamTableIdx()}}] does not match [${{tableIdx}}]`);
-                            }}
-
-                            let removed = cstate.handles.remove(streamEnd.waitableIdx());
-                            if (!removed) {{
-                                 throw new Error(`failed to remove stream end [${{streamEndWaitableIdx}}] waitable obj in component [${{componentIdx}}]`);
-                            }}
-
-                            removed = table.remove(streamEnd.handle());
-                            if (!removed) {{
-                                 throw new Error(`failed to remove stream end with handle [${{streamEnd.handle()}}] from stream table [${{tableIdx}}] in component [${{componentIdx}}]`);
-                            }}
-
-                            return streamEnd;
-                        }}
-
-                        removeStreamEndFromTable(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#removeStreamEndFromTable()] args', args);
-
-                            const {{ tableIdx, streamWaitableIdx }} = args;
-                            if (tableIdx === undefined) {{ throw new Error("missing table idx while removing stream end"); }}
-                            if (streamWaitableIdx === undefined) {{
-                                throw new Error("missing stream end waitable idx while removing stream end");
-                            }}
-
-                            const {{ table, componentIdx }} = {global_stream_table_map}[tableIdx];
-                            if (!table) {{ throw new Error(`missing/invalid table [${{tableIdx}}] while removing stream end`); }}
-
-                            const cstate = {get_or_create_async_state_fn}(componentIdx);
-
-                            const streamEnd = cstate.handles.get(streamWaitableIdx);
-                            if (!streamEnd) {{
-                                throw new Error(`missing stream end (handle [${{streamWaitableIdx}}], table [${{tableIdx}}])`);
-                            }}
-                            const handle = streamEnd.handle();
-
-                            let removed = cstate.handles.remove(streamWaitableIdx);
-                            if (!removed) {{
-                                throw new Error(`failed to remove streamEnd from handles (waitable idx [${{streamWaitableIdx}}]), component [${{componentIdx}}])`);
-                            }}
-
-                            removed = table.remove(handle);
-                            if (!removed) {{
-                                throw new Error(`failed to remove streamEnd from table (handle [${{handle}}]), table [${{tableIdx}}], component [${{componentIdx}}])`);
-                            }}
-
-                            return streamEnd;
-                        }}
-
-                        createFuture(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#createFuture()] args', args);
-                            const {{ tableIdx, elemMeta, hostInjectFn }} = args;
-                            if (tableIdx === undefined) {{ throw new Error("missing table idx while adding future"); }}
-                            if (elemMeta === undefined) {{ throw new Error("missing element metadata while adding future"); }}
-
-                            const {{ table: futureTable, componentIdx }} = {global_future_table_map}[tableIdx];
-                            if (!futureTable) {{
-                                throw new Error(`missing global future table lookup for table [${{tableIdx}}] while creating future`);
-                            }}
-                            if (componentIdx !== this.#componentIdx) {{
-                                throw new Error('component idx mismatch while creating future');
-                            }}
-
-                            const readWaitable = this.createWaitable();
-                            const writeWaitable = this.createWaitable();
-
-                            const future = new {internal_future_class}({{
-                                tableIdx,
-                                componentIdx: this.#componentIdx,
-                                elemMeta,
-                                readWaitable,
-                                writeWaitable,
-                                hostInjectFn,
-                            }});
-                            future.setGlobalFutureMapRep({global_future_map}.insert(future));
-
-                            const writeEnd = future.writeEnd();
-                            writeEnd.setWaitableIdx(this.handles.insert(writeEnd));
-                            writeEnd.setHandle(futureTable.insert(writeEnd));
-                            if (writeEnd.futureTableIdx() !== tableIdx) {{ throw new Error("unexpectedly mismatched future table"); }}
-
-                            const writeEndWaitableIdx = writeEnd.waitableIdx();
-                            const writeEndHandle = writeEnd.handle();
-                            writeWaitable.setTarget(`waitable for future write end (waitable [${{writeEndWaitableIdx}}])`);
-                            writeEnd.setTarget(`future write end (waitable [${{writeEndWaitableIdx}}])`);
-
-                            const readEnd = future.readEnd();
-                            readEnd.setWaitableIdx(this.handles.insert(readEnd));
-                            readEnd.setHandle(futureTable.insert(readEnd));
-                            if (readEnd.futureTableIdx() !== tableIdx) {{ throw new Error("unexpectedly mismatched future table"); }}
-
-                            const readEndWaitableIdx = readEnd.waitableIdx();
-                            const readEndHandle = readEnd.handle();
-                            readWaitable.setTarget(`waitable for read end (waitable [${{readEndWaitableIdx}}])`);
-                            readEnd.setTarget(`future read end (waitable [${{readEndWaitableIdx}}])`);
-
-                            return {{
-                                writeEnd,
-                                writeEndWaitableIdx,
-                                writeEndHandle,
-                                readEndWaitableIdx,
-                                readEndHandle,
-                                readEnd,
-                            }};
-                        }}
-
-                        getFutureEnd(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#getFutureEnd()] args', args);
-                            const {{ tableIdx, futureEndHandle, futureEndWaitableIdx }} = args;
-                            if (tableIdx === undefined) {{
-                                throw new Error('missing table idx while getting future end');
-                            }}
-
-                            const {{ table, componentIdx }} = {global_future_table_map}[tableIdx];
-                            const cstate = {get_or_create_async_state_fn}(componentIdx);
-
-                            let futureEnd;
-                            if (futureEndWaitableIdx !== undefined) {{
-                                futureEnd = cstate.handles.get(futureEndWaitableIdx);
-                            }} else if (futureEndHandle !== undefined) {{
-                                if (!table) {{ throw new Error(`missing/invalid table [${{tableIdx}}] while getting future end`); }}
-                                futureEnd = table.get(futureEndHandle);
-                            }} else {{
-                                throw new TypeError("must specify either waitable idx or handle to retrieve future");
-                            }}
-
-                            if (!futureEnd) {{
-                                throw new Error(`missing future end (tableIdx [${{tableIdx}}], handle [${{futureEndHandle}}], waitableIdx [${{futureEndWaitableIdx}}])`);
-                            }}
-                            if (tableIdx && futureEnd.futureTableIdx() !== tableIdx) {{
-                                throw new Error(`future end table idx [${{futureEnd.futureTableIdx()}}] does not match [${{tableIdx}}]`);
-                            }}
-
-                            return futureEnd;
-                        }}
-
-                        addFutureEndToTable(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#addFutureEndToTable()] args', args);
-                            const {{ tableIdx, futureEnd }} = args;
-                            if (typeof futureEnd === 'number') {{ throw new Error("INSERTING BAD FUTUREEND"); }}
-
-                            let {{ table, componentIdx }} = {global_future_table_map}[tableIdx];
-                            if (componentIdx === undefined || !table) {{
-                                throw new Error(`invalid global future table state for table [${{tableIdx}}]`);
-                            }}
-
-                            const handle = table.insert(futureEnd);
-                            futureEnd.setHandle(handle);
-                            futureEnd.setFutureTableIdx(tableIdx);
-
-                            const cstate = {get_or_create_async_state_fn}(componentIdx);
-                            const waitableIdx = cstate.handles.insert(futureEnd);
-                            futureEnd.setWaitableIdx(waitableIdx);
-
-                            {debug_log_fn}('[{component_async_state_class}#addFutureEndToTable()] added future end', {{
-                                tableIdx,
-                                table,
-                                handle,
-                                futureEnd,
-                                destComponentIdx: componentIdx,
-                            }});
-
-                            return {{ handle, waitableIdx }};
-                        }}
-
-                        removeFutureEndFromTable(args) {{
-                            {debug_log_fn}('[{component_async_state_class}#removeFutureEndFromTable()] args', args);
-                            const {{ tableIdx, futureWaitableIdx }} = args;
-                            if (tableIdx === undefined) {{ throw new Error("missing table idx while removing future end"); }}
-                            if (futureWaitableIdx === undefined) {{
-                                throw new Error("missing future end waitable idx while removing future end");
-                            }}
-
-                            const {{ table, componentIdx }} = {global_future_table_map}[tableIdx];
-                            if (!table) {{ throw new Error(`missing/invalid table [${{tableIdx}}] while removing future end`); }}
-
-                            const cstate = {get_or_create_async_state_fn}(componentIdx);
-
-                            const futureEnd = cstate.handles.get(futureWaitableIdx);
-                            if (!futureEnd) {{
-                                throw new Error(`missing future end (handle [${{futureWaitableIdx}}], table [${{tableIdx}}])`);
-                            }}
-                            const handle = futureEnd.handle();
-
-                            let removed = cstate.handles.remove(futureWaitableIdx);
-                            if (!removed) {{
-                                throw new Error(`failed to remove futureEnd from handles (waitable idx [${{futureWaitableIdx}}]), component [${{componentIdx}}])`);
-                            }}
-
-                            removed = table.remove(handle);
-                            if (!removed) {{
-                                throw new Error(`failed to remove futureEnd from table (handle [${{handle}}]), table [${{tableIdx}}], component [${{componentIdx}}])`);
-                            }}
-
-                            return futureEnd;
-                        }}
-
                     }}
                     "#,
                 ));
